% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_charge_marche.R
\name{func_charge_marche}
\alias{func_charge_marche}
\title{func_charge_marche}
\usage{
func_charge_marche(fichier_prog, connexion, mar_id)
}
\arguments{
\item{fichier_prog}{: fichier xlsx définissant le marché.
Le fichier contient les onglets programmes_types, BPU, cout_run_analytiques
Ces onglets ont a minima les colonnes prévues au dictionnaire de données}

\item{connexion}{: connexion à la BDD}

\item{mar_id}{: identifiant du marché concerné par la programmation dans la table marché

func_charge_marche(fichier_prog, connexion, mar_id)}
}
\value{
Un programme prévisionnel
}
\description{
Fonction pour charger les marchés à partir des fichiers EXCEL
au format imposé et le bancariser dans la bdd
Sont chargés BPU, cout_run_analytiques et programmes_types
}
