% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_charge_ref_sandre_parametres.R
\name{func_charge_referentiel_SANDRE_parametres}
\alias{func_charge_referentiel_SANDRE_parametres}
\title{func_charge_ref_sandre_parametres}
\usage{
func_charge_referentiel_SANDRE_parametres(date_maj = "1950-01-01")
}
\arguments{
\item{date_maj}{: date à partir de laquelle telecharger les mises à jour}
}
\value{
The return value, if any, from executing the function.
}
\description{
Charge le référentiel SANDRE des paramètres
}
