% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_enregistre_dataframe_bdd.R
\name{func_enregistre_dataframe_bdd}
\alias{func_enregistre_dataframe_bdd}
\title{func_enregistre_dataframe_bdd}
\usage{
func_enregistre_dataframe_bdd(
  dataframe_a_enr,
  table_destination,
  schema_destination,
  connexion
)
}
\arguments{
\item{dataframe_a_enr}{Dataframe à enregistrer dans la base.
Les noms de colonnes doivent correspondre aux noms de champs dans la base.}

\item{table_destination}{: nom postgreSQL de la table dans laquelle on souhaite enregistrer}

\item{schema_destination}{: nom postgreSQL du schéma sous lequel enregistrer}

\item{connexion}{: connexion vers la base postgreSQL}
}
\value{
résultat de l'exéctution de la requête.
}
\description{
Fonction pour ajouter un dataframe au contenu d'une table de la base libreSQE
}
