% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_genere_bdc_xlsx.R
\name{func_genere_bdc_xlsx}
\alias{func_genere_bdc_xlsx}
\title{func_genere_bdc_xlsx}
\usage{
func_genere_bdc_xlsx(bdc_id, pre_id, connexion, bdc_type, out_file_name)
}
\arguments{
\item{bdc_id}{: identifiant du bon de commande à éditer}

\item{pre_id}{: identifiant du (des) prestataire(s)}

\item{connexion}{: connexion à la BDD}

\item{bdc_type}{: ficier xlsx, trame du bon de commande à générer}

\item{out_file_name}{: chemin d'enregistrement du fichier}
}
\value{
Un fichier xlsx
}
\description{
Génère un bon de commande au format xlsx
}
