% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_importe_quesu.R
\name{func_importe_quesu}
\alias{func_importe_quesu}
\title{func_importe_quesu}
\usage{
func_importe_quesu(fichier, stations_a_traiter = NULL)
}
\arguments{
\item{fichier}{: chemin d'accès au fichier à ouvrir}

\item{stations_a_traiter}{: paramètre optionnel listant les codes SANDRE des stations à importer}
}
\value{
liste avec 4 éléments : Analyses qui contient la liste des analyses importées,
Echantillon qui contient les caractéristiques de transport des échantillons,
Res_env qui contient les informations environnementales du prélèvement,
}
\description{
lecture de fichiers SANDRE xml aux formats d'échanges QUESU 3 ou 3.1
}
