% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_lit_le_fichier.R
\name{func_lit_le_fichier}
\alias{func_lit_le_fichier}
\title{func_lit_le_fichier}
\usage{
func_lit_le_fichier(fichier_prog, onglet)
}
\arguments{
\item{fichier_prog}{: fichier xlsx définissant le marché.
Le fichier contient les onglets programmes_types, BPU, cout_run_analytiques
Ces onglets ont a minima les colonnes prévues au dictionnaire de données

fonction lecture du fichier}

\item{onglet}{: fichier xlsx définissant le marché.}
}
\value{
Un dataframe contenant les onglets de programmation avec les formats de données
conformes qu dictionnaire de données
}
\description{
Fonction pour charger les onglets marchés à partir des fichiers EXCEL au format imposé
}
