% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_maj_referentiels_sandre.R
\name{func_maj_referentiels_sandre}
\alias{func_maj_referentiels_sandre}
\title{func_maj_referentiels_sandre}
\usage{
func_maj_referentiels_sandre(connexion)
}
\arguments{
\item{connexion}{: connexion à la bdd postgreSQL}
}
\value{
Met à jour la bdd postgreSQL associée à libre SQE en actualisant les tables liées au référentiel SANDRE.
}
\description{
Fonction pour mettre à jour les référentiels SANDRE dans la base de données
}
