% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_test_metier_coherenceO2.R
\name{func_test_metier_coherenceO2}
\alias{func_test_metier_coherenceO2}
\title{func_test_metier_coherenceO2}
\usage{
func_test_metier_coherenceO2(
  O2 = NULL,
  satO2 = NULL,
  temp = NULL,
  export = "code"
)
}
\arguments{
\item{O2}{: dissolved oxygen content in mg O2/L}

\item{satO2}{: oxygen saturation in % (value between 0 and 250, more than 100% = oversaturated)}

\item{temp}{: temperature in °C}

\item{export}{: "code" (default) : return code if correlation is OK or not,
"value" return oxygen saturation calculated for given O2 and temp values.}
}
\value{
The return value, is either a numeric code :
 0 = impossible to do the test
 1 = correlation OK (difference between provide satO2 value and calculated satO2value respect
 satO2 value in [satO2 calculated - 2, sat O2 calculated + 2])
 2 = correlation not exactly respected (satO2 value in [satO2 calculated - 5, sat O2 calculated - 1[ or
 satO2 value in ]satO2 calculated +1, sat O2 calculated + 5] )
 3 = error for sure

or the calculated value of oxygen saturation for given temperature and dissolved O2.
}
\description{
In pure water under 1atm of pressure, the parameters of temperature,
dissolved oxygen content, and oxygen saturation are inherently correlated by definition.
This function test if this is the case.
}
