% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_test_metier_coherenceP.R
\name{func_test_metier_coherenceP}
\alias{func_test_metier_coherenceP}
\title{func_test_metier_coherenceP}
\usage{
func_test_metier_coherenceP(
  Ptot = NULL,
  PO4 = NULL,
  incertPtot = NULL,
  incertPO4 = NULL,
  forceincert = FALSE
)
}
\arguments{
\item{Ptot}{: total phosphorus in mg P/L}

\item{PO4}{: orthophosphates in mg PO4 / L}

\item{incertPtot}{: Measurement uncertainty on Ptot measurement (in % of value) - default = 20 (%)}

\item{incertPO4}{: Measurement uncertainty on PO4 measurement (in % of value) - defaut = 20 (%)}

\item{forceincert}{: boolean. If TRUE, an incorrect value of incertPtot or of incertPO4 will be
automaticaly replaced by 20 (in %)}
}
\value{
The return value, is a numeric code :
 0 = impossible to do the test
 1 = OK PO4<=Ptot without considering uncertainty
 2 = OK only by considering uncertainty
 3 = PO4 measurement to high versus Ptot measurement
}
\description{
test that total phosphorus is greater or equal to orthophosphates
}
