% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_update_table.R
\name{func_update_table}
\alias{func_update_table}
\title{func_update_table}
\usage{
func_update_table(
  dataframe_a_enr,
  cle,
  table_destination,
  schema_destination,
  connexion
)
}
\arguments{
\item{dataframe_a_enr}{dataframe à enregistrer dans la base. Ses noms de colonnes doivent
correspondre aux noms de champs de la bdd}

\item{cle}{: nom commun de la colonne de data.frame et de la cle primaire de la table de destination}

\item{table_destination}{: nom postgreSQL de la table dans laquelle on souhaite enregistrer}

\item{schema_destination}{: nom postgreSQL du schéma sous lequel enregistrer}

\item{connexion}{: connexion vers la base postgreSQL



Les noms de colonnes doivent correspondre aux noms de champs dans la base.}
}
\value{
The return value, if any, from executing the function.

résultat de l'exécution de la requête.
}
\description{
Fonction pour mettre à jour dans la base postgresql une table
}
