% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ss_func_extrat_from_xml.R
\name{ss_func_extrat_from_xml}
\alias{ss_func_extrat_from_xml}
\title{ss_func_extrat_from_xml}
\usage{
ss_func_extrat_from_xml(xml_nodes, values = NULL, scheme = NULL, count = NULL)
}
\arguments{
\item{xml_nodes}{: nodes to treat}

\item{values}{: list of values to insert in result(xml tag names and xml tag path if necessary)}

\item{scheme}{: list of xml attribute to insert in result (xml tag names and xml tag path if necessary)}

\item{count}{: list of xml tag to count

syntax for the lists values and count:
- First element and following ones = xml tag path,
- last element = xml tag name

syntax for scheme :
list(c(list(path), scheme_name))}
}
\value{
The function return a dataframe with required data as column name
}
\description{
Function which convert xml to dataframe.
}
