% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_ajoute_nom_sandre.R
\name{func_ajoute_nom_sandre}
\alias{func_ajoute_nom_sandre}
\title{func_ajoute_nom_sandre}
\usage{
func_ajoute_nom_sandre(connexion = connexion, code, out = "nom_parametre")
}
\arguments{
\item{connexion}{: DBI connexion to active LibreSQE database}

\item{code}{: a vector containing the code to convert}
}
\value{
vector of same length as input with corresponding name.
If code is missing, value return is NA
}
\description{
A function to add name according to french reference SANDRE corresponding to a code
}
