% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_charge_prog_annuelle.R
\name{func_charge_prog_annuelle}
\alias{func_charge_prog_annuelle}
\title{func_charge_prog_annuelle}
\usage{
func_charge_prog_annuelle(fichier_prog, connexion, mar_id, annee)
}
\arguments{
\item{fichier_prog}{: fichier xlsx définissant le marché.
Le fichier contient les onglets programme_annuel et calendrier
Ces onglets ont a minima les colonnes prévues au dictionnaire de données}

\item{connexion}{: connexion à la BDD}

\item{mar_id}{: identifiant du marché concerné par la programmation dans la table marché}

\item{annee}{: année civile de la programmation saisie

func_charge_marche(fichier_prog, connexion, mar_id)}
}
\value{
Un programme prévisionnel
}
\description{
Fonction pour charger les programmations annuelles à partir des fichiers EXCEL
au format imposé et le bancariser dans la bdd
sont chargés programme_annuel et calendrier
}
