% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_func_test_metier_coherenceO2.R
\name{func_test_metier_coherenceO2}
\alias{func_test_metier_coherenceO2}
\title{func_test_metier_coherenceO2}
\usage{
func_test_metier_coherenceO2(O2 = NULL, satO2 = NULL, temp = NULL)
}
\arguments{
\item{O2}{: dissolved oxygen content in mg O2/L}

\item{satO2}{: oxygen saturation in % (value between 0 and 250, more than 100% = oversaturated)}

\item{temp}{: temperature in °C}
}
\value{
The return value, is a numeric code :
 0 = impossible to do the test
 1 = correlation OK (less than 1% of difference between provide satO2 value and calculated satO2value)
 2 = correlation not exactly respected (less than 5% of difference between provide satO2 value and calculated satO2value)
 3 = error for sure
}
\description{
In pure water under 1atm of pressure, the parameters of temperature,
dissolved oxygen content, and oxygen saturation are inherently correlated by definition.
This function test if this is the case.
}
